# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "SIMANTIC - Change & Configuration API",
        "description": "API untuk CMDB, Change Management, Maintenance & Patch Management",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost",
            "description": "Local development server"
        }
    ],
    "paths": {
...
        "/api/v1/rfc": {
            "get": {
                "tags": [
                    "RFC"
                ],
                "summary": "Daftar RFC milik OPD user login",
                "description": "GET /api/v1/rfc\nMenampilkan daftar RFC milik OPD user login (dengan pagination).",
                "operationId": "7478032a40e807e32f132a4f6077fe0e",
                "parameters": [
                    {
                        "name": "status",
                        "in": "query",
                        "description": "Filter status RFC (optional)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "submitted"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Jumlah data per halaman, default 15",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 15
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Berhasil mengambil daftar RFC",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "title": {
                                                        "type": "string",
                                                        "example": "Upgrade Database SIM Pegawai"
                                                    },
                                                    "category": {
                                                        "type": "string",
                                                        "example": "normal"
                                                    },
                                                    "urgency": {
                                                        "type": "string",
                                                        "example": "high"
                                                    },
                                                    "priority": {
                                                        "type": "string",
                                                        "example": "P2"
                                                    },
                                                    "status": {
                                                        "type": "string",
                                                        "example": "submitted"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "meta": {
                                            "properties": {
                                                "current_page": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "last_page": {
                                                    "type": "integer",
                                                    "example": 5
                                                },
                                                "per_page": {
                                                    "type": "integer",
                                                    "example": 15
                                                },
                                                "total": {
                                                    "type": "integer",
                                                    "example": 73
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "RFC"
                ],
                "summary": "Buat RFC baru",
                "description": "POST /api/v1/rfc\nMembuat RFC baru dan menghubungkan dengan CI terkait + auto assessment.",
                "operationId": "a499ce14e083680fe755781dce428af1",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title",
                                    "category",
                                    "urgency",
                                    "priority",
                                    "ci_ids"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "maxLength": 255,
                                        "example": "Upgrade Database SIM Pegawai"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Perlu upgrade versi 13 untuk compliance."
                                    },
                                    "category": {
                                        "type": "string",
                                        "enum": [
                                            "normal",
                                            "standard",
                                            "emergency"
                                        ],
                                        "example": "normal"
                                    },
                                    "urgency": {
                                        "type": "string",
                                        "enum": [
                                            "low",
                                            "medium",
                                            "high",
                                            "critical"
                                        ],
                                        "example": "high"
                                    },
                                    "priority": {
                                        "type": "string",
                                        "enum": [
                                            "P4",
                                            "P3",
                                            "P2",
                                            "P1"
                                        ],
                                        "example": "P2"
                                    },
                                    "ci_ids": {
                                        "description": "Daftar ID Configuration Item yang terdampak",
                                        "type": "array",
                                        "items": {
                                            "type": "integer",
                                            "example": 101
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "RFC berhasil dibuat",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "title": {
                                            "type": "string"
                                        },
                                        "category": {
                                            "type": "string",
                                            "example": "normal"
                                        },
                                        "urgency": {
                                            "type": "string",
                                            "example": "high"
                                        },
                                        "priority": {
                                            "type": "string",
                                            "example": "P2"
                                        },
                                        "status": {
                                            "type": "string",
                                            "example": "submitted"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validasi gagal atau user belum memiliki OPD",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "User has no OPD assigned"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/rfc/{id}": {
            "get": {
                "tags": [
                    "RFC"
                ],
                "summary": "Detail satu RFC (lengkap)",
                "description": "GET /api/v1/rfc/{rfc}\nMenampilkan detail RFC lengkap (dengan guard OPD).",
                "operationId": "7b1bd1f0342b681d5103cb13eec1c932",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID RFC",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detail RFC",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "title": {
                                            "type": "string",
                                            "example": "Upgrade Database SIM Pegawai"
                                        },
                                        "description": {
                                            "type": "string"
                                        },
                                        "category": {
                                            "type": "string",
                                            "example": "normal"
                                        },
                                        "urgency": {
                                            "type": "string",
                                            "example": "high"
                                        },
                                        "priority": {
                                            "type": "string",
                                            "example": "P2"
                                        },
                                        "status": {
                                            "type": "string",
                                            "example": "approved"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Tidak boleh melihat RFC OPD lain"
                    },
                    "404": {
                        "description": "RFC tidak ditemukan"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}